# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /masters/cells:
    get:
      operationId: listCellMasters
      tags: [masters]
      parameters:
        - $ref: "#/components/parameters/SearchParam"
        - $ref: "#/components/parameters/StatusParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: List of cell masters
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/MasterListResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/CellMaster"
    post:
      operationId: createCellMaster
      tags: [masters]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CellMasterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CellMaster"
  /masters/cells/{id}:
    get:
      operationId: getCellMaster
      tags: [masters]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Single cell master
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CellMaster"
    patch:
      operationId: updateCellMaster
      tags: [masters]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CellMasterUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CellMaster"
...
  /cells/lots:
    get:
      operationId: listCellLots
      tags: [cells]
      summary: List cell receiving lots
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/SearchParam"
      responses:
        "200":
          description: List of lots
          content:
            application/json:
              schema:
                type: object
                required: [items, meta]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/CellLot"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
    post:
      operationId: createCellLot
      tags: [cells]
      summary: Receive a new cell lot (auto-generates individual cell records)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CellLotInput"
      responses:
        "201":
          description: Created lot with stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CellLotDetail"
  /cells/lots/{id}:
    get:
      operationId: getCellLot
      tags: [cells]
      summary: Get cell lot detail with cell stats
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Lot detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CellLotDetail"
        "404":
          description: Not found

  # ── Cell Grading — Matches ────────────────────────────────────────────────
...
  /cells:
    get:
      operationId: listCells
      tags: [cells]
      summary: List individual cells with filters
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/SearchParam"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [received, grading, approved, rejected, quarantine, reserved, allocated]
        - name: grade
          in: query
          required: false
          schema:
            type: string
            enum: [A, B, C, reject]
        - name: lotId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of cells
          content:
            application/json:
              schema:
                type: object
                required: [items, meta]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Cell"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
  /cells/{id}:
    get:
      operationId: getCell
      tags: [cells]
      summary: Get cell detail with traceability
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Cell detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CellDetail"
        "404":
          description: Not found
  /cells/{id}/grade:
    post:
      operationId: gradeCell
      tags: [cells]
      summary: Record grading measurements and auto-calculate grade
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CellGradeInput"
      responses:
        "200":
          description: Graded cell
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cell"
        "400":
          description: Cell not in gradeable state

  # ── Cell Grading — Reports ────────────────────────────────────────────────
...
  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Login successful — sets httpOnly cookie
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: authMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Authenticated user info
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: "#/components/schemas/AuthUser" }
        "401":
          description: Not authenticated

  # ─── Director Dashboard ───────────────────────────────────────────────────
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SearchParam:
      name: search
      in: query
      required: false
      schema:
        type: string
    StatusParam:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum: [active, inactive]
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        default: 25
  schemas:
...
    MasterStatus:
      type: string
      enum: [active, inactive]
...
    Attachment:
      type: object
      required: [id, name, url, fileType, uploadedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        url: { type: string }
        fileType: { type: string }
        uploadedAt: { type: string, format: date-time }
    MasterCommon:
      type: object
      required: [id, code, name, status, revision_number, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/MasterStatus" }
        revision_number: { type: integer }
        effective_date: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
        created_by: { type: ["string", "null"], format: uuid }
        created_at: { type: string, format: date-time }
        updated_by: { type: ["string", "null"], format: uuid }
        updated_at: { type: string, format: date-time }
    MasterCommonInput:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        effective_date: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
    MasterCommonUpdate:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        effective_date: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
    MasterListMeta:
      type: object
      required: [total, page, pageSize, totalPages]
      properties:
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
        totalPages: { type: integer }
    MasterListResponse:
      type: object
      required: [items, meta]
      properties:
        meta: { $ref: "#/components/schemas/MasterListMeta" }

    # Product Master
...
    CellMaster:
      allOf:
        - $ref: "#/components/schemas/MasterCommon"
        - type: object
          properties:
            manufacturer: { type: string }
            model: { type: string }
            chemistry: { type: string }
            capacity_mah: { type: number }
            nominal_voltage_v: { type: number }
            max_voltage_v: { type: number }
            min_voltage_v: { type: number }
            weight_g: { type: number }
            dimensions: { type: string }
            internal_resistance_spec_mohm: { type: number }
            cycle_life: { type: integer }
            datasheet_url: { type: ["string", "null"] }
            approved_supplier: { type: ["string", "null"] }
    CellMasterInput:
      allOf:
        - $ref: "#/components/schemas/MasterCommonInput"
        - type: object
          required: [manufacturer, model, chemistry, capacity_mah, nominal_voltage_v, max_voltage_v, min_voltage_v, weight_g, dimensions, internal_resistance_spec_mohm, cycle_life]
          properties:
            manufacturer: { type: string }
            model: { type: string }
            chemistry: { type: string }
            capacity_mah: { type: number }
            nominal_voltage_v: { type: number }
            max_voltage_v: { type: number }
            min_voltage_v: { type: number }
            weight_g: { type: number }
            dimensions: { type: string }
            internal_resistance_spec_mohm: { type: number }
            cycle_life: { type: integer }
            datasheet_url: { type: ["string", "null"] }
            approved_supplier: { type: ["string", "null"] }
    CellMasterUpdate:
      allOf:
        - $ref: "#/components/schemas/MasterCommonUpdate"
        - type: object
          properties:
            manufacturer: { type: string }
            model: { type: string }
            chemistry: { type: string }
            capacity_mah: { type: number }
            nominal_voltage_v: { type: number }
            max_voltage_v: { type: number }
            min_voltage_v: { type: number }
            weight_g: { type: number }
            dimensions: { type: string }
            internal_resistance_spec_mohm: { type: number }
            cycle_life: { type: integer }
            datasheet_url: { type: ["string", "null"] }
            approved_supplier: { type: ["string", "null"] }

    # BMS Master
...
    PaginationMeta:
      type: object
      required: [total, page, pageSize, totalPages]
      properties:
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
        totalPages: { type: integer }
...
    CellLot:
      type: object
      required: [id, supplier, manufacturer, cellModel, cellChemistry, nominalCapacityAh, lotNumber, dateReceived, quantityReceived, receivedBy, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        supplier: { type: string }
        manufacturer: { type: string }
        cellModel: { type: string }
        cellChemistry: { type: string }
        nominalCapacityAh: { type: number }
        lotNumber: { type: string }
        invoiceNumber: { type: ["string", "null"] }
        dateReceived: { type: string }
        quantityReceived: { type: integer }
        receivedBy: { type: string }
        remarks: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CellLotDetail:
      allOf:
        - $ref: "#/components/schemas/CellLot"
        - type: object
          properties:
            stats:
              type: object
              properties:
                total: { type: integer }
                received: { type: integer }
                approved: { type: integer }
                rejected: { type: integer }
                quarantine: { type: integer }
                reserved: { type: integer }
                allocated: { type: integer }
    CellLotInput:
      type: object
      required: [supplier, manufacturer, cellModel, nominalCapacityAh, lotNumber, dateReceived, quantityReceived, receivedBy]
      properties:
        supplier: { type: string }
        manufacturer: { type: string }
        cellModel: { type: string }
        cellChemistry: { type: string, default: LiFePO4 }
        nominalCapacityAh: { type: number }
        lotNumber: { type: string }
        invoiceNumber: { type: ["string", "null"] }
        dateReceived: { type: string }
        quantityReceived: { type: integer, minimum: 1 }
        receivedBy: { type: string }
        remarks: { type: ["string", "null"] }
    Cell:
      type: object
      required: [id, cellId, lotId, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        cellId: { type: string }
        lotId: { type: string, format: uuid }
        status:
          type: string
          enum: [received, grading, approved, rejected, quarantine, reserved, allocated]
        grade:
          type: ["string", "null"]
          enum: [A, B, C, reject, "null"]
        voltageV: { type: ["number", "null"] }
        capacityAh: { type: ["number", "null"] }
        internalResistanceMohm: { type: ["number", "null"] }
        temperatureC: { type: ["number", "null"] }
        gradingMachineId: { type: ["string", "null"] }
        gradedBy: { type: ["string", "null"] }
        gradedAt: { type: ["string", "null"], format: date-time }
        gradingNotes: { type: ["string", "null"] }
        matchId: { type: ["string", "null"], format: uuid }
        allocationOrderId: { type: ["string", "null"], format: uuid }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CellDetail:
      allOf:
        - $ref: "#/components/schemas/Cell"
        - type: object
          properties:
            lot:
              $ref: "#/components/schemas/CellLot"
            usedInOrders:
              type: array
              items:
                type: object
                properties:
                  orderId: { type: string, format: uuid }
                  orderNumber: { type: string }
                  batteryNumber: { type: string }
    CellGradeInput:
      type: object
      required: [voltageV, capacityAh, internalResistanceMohm, gradedBy]
      properties:
        voltageV: { type: number }
        capacityAh: { type: number }
        internalResistanceMohm: { type: number }
        temperatureC: { type: ["number", "null"] }
        gradingMachineId: { type: ["string", "null"] }
        gradedBy: { type: string }
        gradingNotes: { type: ["string", "null"] }
        overrideStatus:
          type: ["string", "null"]
          enum: [approved, rejected, quarantine, "null"]
...
    AuthUser:
      type: object
      required: [userId, email, name, role]
      properties:
        userId: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        role:
          type: string
          enum: [director, supervisor, operator, viewer]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
    LoginResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/AuthUser" }

    # ─── Director Dashboard schema ────────────────────────────────────────────
...